move MakeSignature test
